refactor: reorganize layout module into separate folder
